pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "tinytapeout-cli"
dynamic = ["version"]
description = "Tiny Tapeout CLI - Design, test, and harden ASIC projects"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Electronic Design Automation (EDA)",
]
dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "PyYAML>=6.0",
    "requests>=2.28",
    "packaging",
]

[project.scripts]
tt = "tinytapeout.cli.app:cli"

[project.optional-dependencies]
gds = ["klayout>=0.29,<0.30", "gdstk", "CairoSVG", "numpy<2", "matplotlib", "GitPython"]
fpga = ["yowasp-yosys", "mpremote", "configupdater"]
all = ["tinytapeout-cli[gds,fpga]"]
dev = ["pytest>=7.0", "pre-commit"]

[tool.hatch.version]
source = "vcs"
fallback-version = "0.0.0-dev"

[tool.hatch.build.targets.wheel]
packages = ["src/tinytapeout"]

[tool.hatch.envs.default]
dependencies = ["pytest>=7.0", "pre-commit"]

[tool.hatch.envs.default.scripts]
tt = "python -m tinytapeout {args}"

[tool.ruff]
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.pytest.ini_options]
testpaths = ["tests"]
